requirements.txt
asgiref
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
pytz==2019.3
sqlparse==0.3.0
djangorestframework-simplejwt==5.0.0
django-filter==21.1